finished basic functionality

+++ src/assets/DeleteComment.jsx
@@ -2,20 +2,32 @@ import React, { useState } from "react";
 import axios from "axios";
 import "./CSS/Comments.css";
 import UrlBase from "./UrlBase";
+import { useContext } from "react";
+import { UserContext } from "./User";
 
-const CommentActions = ({ comment_id, onDelete }) => {
+const CommentActions = ({ comment_id, author, onDelete }) => {
   const [confirmDelete, setConfirmDelete] = useState(false);
+  const [errMsg, setErrMsg] = useState(false);
+  const [isError, setIsError] = useState(false);
+  const { user } = useContext(UserContext);
 
   const handleDelete = () => {
     if (confirmDelete) {
-      axios
-        .delete(`${UrlBase}comments/${comment_id}`)
-        .then(() => {
-          onDelete(comment_id);
-        })
-        .catch((err) => {
-          console.log("Error deleting comment", err);
-        });
+      if (author === user) {
+        axios
+          .delete(`${UrlBase}comments/${comment_id}`)
+          .then(() => {
+            onDelete(comment_id);
+          })
+          .catch((err) => {
+            console.log("Error deleting comment", err);
+            setIsError(true);
+            setErrMsg(err);
+          });
+      } else {
+        setIsError(true);
+        setErrMsg("Only the author can delete this comment.");
+      }
     } else {
       setConfirmDelete(true);
     }
@@ -23,7 +35,10 @@ const CommentActions = ({ comment_id, onDelete }) => {
 
   const handleCancelDelete = () => {
     setConfirmDelete(false);
+    setIsError(false)
   };
+
+  if (isError) return <p>{`${errMsg}`}</p>;
 
   return (
     <div className="comment-actions" key={comment_id}>
